fix: port all 253 scripts to Unity 2022.3.45f1 API compatibility
- Replace linearVelocity→velocity, linearDamping→drag across 22 files (94 occurrences)
- Fix CS0070 event invocation from subclasses (BossBase, PlayerAnimator)
- Fix TrackEvent tuple→Dictionary syntax in CurrencyManager, DailyRewardSystem
- Add missing asmdef refs (RenderPipelines.Core, Networking.Transport)
- Downgrade 2D packages to 2022.3 compatible versions (2d.common@8.0.2)
- Fix RequireComponent ordering in PrefabFactory (Animator before PlayerAnimator)
- Add CLI entry point RunFullSetupFromCommandLine to ProjectBootstrapper
- Various type/access fixes (PlayerHealth.Respawn, DailyRewardUI Vector2, etc.)

+++ Assets/Scripts/Editor/AndroidBuildHelper.cs
@@ -244,7 +244,7 @@ public class AndroidBuildHelper : EditorWindow
 
         if (report.summary.result == BuildResult.Succeeded)
         {
-            long sizeBytes = report.summary.totalSize;
+            long sizeBytes = (long)report.summary.totalSize;
             string sizeMB = (sizeBytes / (1024f * 1024f)).ToString("F1");
             EditorUtility.DisplayDialog("构建成功",
                 $"构建完成！\n路径：{outputPath}\n大小：{sizeMB} MB\n耗时：{report.summary.totalTime.TotalSeconds:F0}秒",
